chore: move TestQueryExecutorUser to util.dart
